pyproject.toml
[tool.poetry]
name = "accueil"
version = "1.0.0"
description = "simple interface for odoo annotating members attendance"
authors = ["romain Viry <rom88.viry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^23.2.0"
sanic = "^23.12.1"
sanic-ext = "^23.12.0"
requests = "^2.32.3"
sanic-testing = "^24.6.0"
erppeek = "^1.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "==6.2.5"
pytest-cov = "^3.0.0"
commitizen = "^2.27.1"
yamllint = "^1.35.1"
ruff = "^0.2.2"
safety = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--no-header -l --cov accueil -m 'not running_server'"

testpaths = [
    "tests"
]

[tool.coverage.run]
omit = [
    "tests/*",
]
